requirements.txt
arrow==1.3.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
ics==0.7.2
idna==3.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.4
six==1.17.0
TatSu==5.13.1
types-python-dateutil==2.9.0.20250516
urllib3==2.4.0
